requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0
pytest>=7.4.0
pytest-cov>=4.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas


